curioweb: Move chain state from SSE to rpc
