+++ curiosrc/web/api/routes.go
@@ -6,12 +6,12 @@ import (
 
 	"github.com/filecoin-project/lotus/cmd/curio/deps"
 	"github.com/filecoin-project/lotus/curiosrc/web/api/config"
-	"github.com/filecoin-project/lotus/curiosrc/web/api/debug"
 	"github.com/filecoin-project/lotus/curiosrc/web/api/sector"
+	"github.com/filecoin-project/lotus/curiosrc/web/api/webrpc"
 )
 
 func Routes(r *mux.Router, deps *deps.Deps) {
-	debug.Routes(r.PathPrefix("/debug").Subrouter(), deps)
+	webrpc.Routes(r.PathPrefix("/webrpc").Subrouter(), deps)
 	config.Routes(r.PathPrefix("/config").Subrouter(), deps)
 	sector.Routes(r.PathPrefix("/sector").Subrouter(), deps)
 }
