+++ curiosrc/web/api/webrpc/routes.go
@@ -2,11 +2,17 @@ package webrpc
 
 import (
 	"context"
+
+	"github.com/gorilla/mux"
+	logging "github.com/ipfs/go-log/v2"
+
 	"github.com/filecoin-project/go-jsonrpc"
+
 	"github.com/filecoin-project/lotus/build"
 	"github.com/filecoin-project/lotus/cmd/curio/deps"
-	"github.com/gorilla/mux"
 )
+
+var log = logging.Logger("webrpc")
 
 type WebRPC struct {
 	deps *deps.Deps
@@ -14,6 +20,10 @@ type WebRPC struct {
 
 func (a *WebRPC) Version(context.Context) (string, error) {
 	return build.UserVersion(), nil
+}
+
+func (a *WebRPC) BlockDelaySecs(context.Context) (uint64, error) {
+	return build.BlockDelaySecs, nil
 }
 
 func Routes(r *mux.Router, deps *deps.Deps) {
