+++ curiosrc/web/srv.go
@@ -4,7 +4,6 @@ package web
 import (
 	"context"
 	"embed"
-	"github.com/filecoin-project/lotus/curiosrc/web/webrpc"
 	"io"
 	"io/fs"
 	"net"
@@ -39,7 +38,6 @@ func GetSrv(ctx context.Context, deps *deps.Deps) (*http.Server, error) {
 		return nil, err
 	}
 	api.Routes(mx.PathPrefix("/api").Subrouter(), deps)
-	webrpc.Routes(mx.PathPrefix("/webrpc").Subrouter(), deps)
 
 	var static fs.FS = static
 	if webDev {
